+++ charts/gardener-extension-shoot-lakom-service/templates/clusterfilters-logging.yaml
@@ -13,6 +13,6 @@ spec:
   filters:
   - parser:
       keyName: log
-      parser: extension-shoot-lakom-service-lakom-parser
+      parser: extension-shoot-lakom-service-lakom-parser-{{ include "name" . }}
       reserveData: true
   match: kubernetes.*extension-shoot-lakom-service*lakom*
