+++ charts/gardener-extension-shoot-lakom-service/templates/clusterparsers-logging.yaml
@@ -6,7 +6,7 @@
 apiVersion: fluentbit.fluent.io/v1alpha2
 kind: ClusterParser
 metadata:
-  name: extension-shoot-lakom-service-lakom-parser
+  name: extension-shoot-lakom-service-lakom-parser-{{ include "name" . }}
   labels:
     fluentbit.gardener/type: "seed"
 spec:
