+++ local-setup/operator/extension.yaml
@@ -20,6 +20,8 @@ spec:
         ociRepository:
           ref: local-skaffold/gardener-extension-shoot-lakom-service/charts/extension
       policy: Always
+      values: {}
+      runtimeClusterValues: {}
   resources:
   - kind: Extension
     type: shoot-lakom-service
