+++ local-setup/operator/kustomization.yaml
@@ -26,6 +26,20 @@ patches:
               MFkwEwYHKoZIzj0CAQYIKoZIzj0DAQcDQgAEAKZef+D07q1AyXxf8xw3L23TgDSq
               Scitl0xDpdLGpODUZXG6LbWY3QXS6yhCt0MZH6+Jf5l/8daO7HDRRvbQMg==
               -----END PUBLIC KEY-----
+    - op: add
+      path: /spec/deployment/extension/runtimeClusterValues/controllers
+      value:
+        allowUntrustedImages: true
+        allowInsecureRegistries: true
+        cosignPublicKeys:
+          publicKeys:
+          - name: example-01
+            algorithm: RSASSA-PKCS1-v1_5-SHA256
+            key: |-
+              -----BEGIN PUBLIC KEY-----
+              MFkwEwYHKoZIzj0CAQYIKoZIzj0DAQcDQgAEAKZef+D07q1AyXxf8xw3L23TgDSq
+              Scitl0xDpdLGpODUZXG6LbWY3QXS6yhCt0MZH6+Jf5l/8daO7HDRRvbQMg==
+              -----END PUBLIC KEY-----
   target:
     group: operator.gardener.cloud
     kind: Extension
