+++ pkg/controller/healthcheck/add.go
@@ -32,6 +32,23 @@ var (
 // RegisterHealthChecks registers health checks for each extension resource
 // HealthChecks are grouped by extension (e.g worker), extension.type (e.g aws) and  Health Check Type (e.g SystemComponentsHealthy)
 func RegisterHealthChecks(mgr manager.Manager, opts healthcheck.DefaultAddArgs) error {
+	managedResourceNames := []string{constants.ManagedResourceNamesSeed}
+
+	if len(opts.ExtensionClasses) == 1 && opts.ExtensionClasses[0] == extensionsv1alpha1.ExtensionClassGarden {
+		managedResourceNames = []string{
+			constants.ManagedResourceNamesGardenRuntime,
+			constants.ManagedResourceNamesGardenVirtual,
+		}
+	}
+
+	healthChecks := make([]healthcheck.ConditionTypeToHealthCheck, 0, len(managedResourceNames))
+	for _, managedResourceName := range managedResourceNames {
+		healthChecks = append(healthChecks, healthcheck.ConditionTypeToHealthCheck{
+			ConditionType: string(gardencorev1beta1.ShootControlPlaneHealthy),
+			HealthCheck:   general.CheckManagedResource(managedResourceName),
+		})
+	}
+
 	return healthcheck.DefaultRegistration(
 		constants.ExtensionType,
 		extensionsv1alpha1.SchemeGroupVersion.WithKind(extensionsv1alpha1.ExtensionResource),
@@ -40,15 +57,30 @@ func RegisterHealthChecks(mgr manager.Manager, opts healthcheck.DefaultAddArgs) 
 		mgr,
 		opts,
 		nil,
-		[]healthcheck.ConditionTypeToHealthCheck{
-			// Never register health checks for `managedresource.spec.class==nil` (ManagedResources installing resources in the shoot cluster) here as it is done by gardenlet, see https://github.com/gardener/gardener/blob/v1.71.3/docs/extensions/healthcheck-library.md?plain=1#L99
-			{
-				ConditionType: string(gardencorev1beta1.ShootControlPlaneHealthy),
-				HealthCheck:   general.CheckManagedResource(constants.ManagedResourceNamesSeed),
-			},
-		},
+		healthChecks,
 		sets.New[gardencorev1beta1.ConditionType]())
 }
+
+// func managedResourceHealthChecks(extensionClasses []extensionsv1alpha1.ExtensionClass) []healthcheck.ConditionTypeToHealthCheck {
+// 	managedResourceNames := []string{constants.ManagedResourceNamesSeed}
+
+// 	if len(extensionClasses) == 1 && extensionClasses[0] == extensionsv1alpha1.ExtensionClassGarden {
+// 		managedResourceNames = []string{
+// 			constants.ManagedResourceNamesGardenRuntime,
+// 			constants.ManagedResourceNamesGardenVirtual,
+// 		}
+// 	}
+
+// 	healthChecks := make([]healthcheck.ConditionTypeToHealthCheck, 0, len(managedResourceNames))
+// 	for _, managedResourceName := range managedResourceNames {
+// 		healthChecks = append(healthChecks, healthcheck.ConditionTypeToHealthCheck{
+// 			ConditionType: string(gardencorev1beta1.ShootControlPlaneHealthy),
+// 			HealthCheck:   general.CheckManagedResource(managedResourceName),
+// 		})
+// 	}
+
+// 	return healthChecks
+// }
 
 // AddToManager adds a controller with the default Options.
 func AddToManager(_ context.Context, mgr manager.Manager) error {
